Add local API module for fabric
This module follows the fabric API so that chttpd can pick between the
two implementations. This allows chttpd to expose individual shard files
directly over the clustered HTTP port. Given that best practices dictate
that 5986 be bound to 127.0.0.1 this allows more advanced reading
directly from shards.

Currently the write operations have been disabled and will through a 501
if any function is required to modify the database. We may want to
investigate allowing for writes in the future. However we'll need to
make sure that individual shard files are able to know their own shard
range so they are capable of rejecting writes that are to the wrong
shard range. Otherwise a user could very easily get their clustered
database into a broken state where docs are on the wrong shards.

COUCHDB-2791